+++ Changelog.md
@@ -30,6 +30,8 @@
 - Don't overwrite layers with exposure overrides by changing the global exposure config
 - Fix loading of .nanodlp sliced files
 - Multi-threaded nanodlp loading
+- Multi-threaded output layer encoding (mainly for nanodlp where encoding is quite expensive)
+- Don't block the main thread while encoding layers
 
 ## v0.9.2 &mdash; August 30th, 2026